+++ src/main/java/com/alarmcontrol/server/data/repositories/EmployeeSkillRepository.java
@@ -7,5 +7,6 @@ import org.springframework.data.repository.CrudRepository;
 public interface EmployeeSkillRepository extends CrudRepository<EmployeeSkill, Long> {
   List<EmployeeSkill> findByEmployeeIdAndSkillId(Long employeeId, Long skillId);
   List<EmployeeSkill> findByEmployeeId(Long employeeId);
+  List<EmployeeSkill> findByEmployeeIdIn(Iterable<Long> employeeIds);
   long countBySkillId(Long skillId);
 }